+++ core/src/main/scala/coulomb/ops/standard/all.scala
@@ -25,4 +25,6 @@ export tquot.given
 export pow.given
 export tpow.given
 export ord.given
-export simplify.given
+
+// will be exported via policies
+export coulomb.ops.simplification.standard.given
